fix: drop ambiguous ModeSetting ctor; add HUDMod no-arg getX/Y/Width/Height; widen Double→int/float casts in modules
- ModeSetting: removing the (name,default,modes...) overload that
  was ambiguous against (name,desc,default,modes...)
- HUDMod: add no-arg getX(), getY(), getWidth(), getHeight() so
  modules don't have to thread screen dims through every call
- Modules: '(int) X.getValue()' is illegal in Java when X.getValue()
  returns Double (cast cannot do unbox + narrowing in one step);
  sed-replaced with '(int)(double) X.getValue()'. Same for (float).

+++ src/main/java/net/glacierclient/modules/advanced/atmosphere/CinematicCameraTweak.java
@@ -38,12 +38,12 @@ public class CinematicCameraTweak extends GlacierMod {
         float factor = (float) (1.0 / smoothness.getValue());
         if (horizontal.getValue()) {
             float yawDiff = mc.player.getYaw() - smoothYaw;
-            float capped = Math.max(-(float) maxSpeed.getValue(), Math.min((float) maxSpeed.getValue(), yawDiff));
+            float capped = Math.max(-(float)(double) maxSpeed.getValue(), Math.min((float)(double) maxSpeed.getValue(), yawDiff));
             smoothYaw += capped * factor;
         }
         if (vertical.getValue()) {
             float pitchDiff = mc.player.getPitch() - smoothPitch;
-            float capped = Math.max(-(float) maxSpeed.getValue(), Math.min((float) maxSpeed.getValue(), pitchDiff));
+            float capped = Math.max(-(float)(double) maxSpeed.getValue(), Math.min((float)(double) maxSpeed.getValue(), pitchDiff));
             smoothPitch += capped * factor;
         }
     }
